minor bugfix

+++ trunk/httptServerDirectEvilB.cc
@@ -55,7 +55,7 @@ std::string httptServerDirectEvilB::generateBody()
 	for( int i=0; i<numResources; i++ )
 	{		
 		rndDelay = 10.0+uniform(0,2.0);
-		refSize = uniform(500,1000); // The random size represents a random reference string length
+		refSize = (int)uniform(500,1000); // The random size represents a random reference string length
 		sprintf(tempBuf, "TEXT%.4d.txt;%s;%f;%s;%d\n", i, "www.good.com", rndDelay, "TRUE", refSize);
 		result.append(tempBuf);
 	}
